backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pymupdf>=1.23.0
pydantic>=2.0.0
httpx>=0.26.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
supabase>=2.0.0
